FastAPI-GenAI/hello-fastapi/pyproject.toml
[tool.poetry]
name = "todos"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
mypy = "^1.12.0"
sqlmodel = "^0.0.22"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"


[tool.poetry.scripts]
start= "todos.sqlModel:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
